feat: Refactor file upload structure and implement new validation schemas
